requirements.txt
asgiref==3.2.3
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.1
django-allauth==0.40.0
django-cors-headers==3.2.0
django-extensions==2.2.5
django-rest-auth==0.9.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
greenlet==0.4.15
gunicorn==20.0.4
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
numpy==1.17.4
oauthlib==3.1.0
parso==0.5.1
psycopg2==2.8.4
PyJWT==1.7.1
pylint==2.4.4
pynvim==0.4.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
regex==2019.11.1
requests==2.22.0
requests-oauthlib==1.3.0
scikit-learn==0.22
scipy==1.3.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
